Broadly refactor search tests
- move test_elastic.py to test_search/test_elastic.py
- factor out test_search/test_views.py from test_views.py
