+++ tests/test_search/test_views.py
@@ -1,0 +1,229 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+from nose.tools import *  # noqa PEP8 asserts
+
+from tests.base import OsfTestCase
+from tests.factories import (
+    ApiOAuth2ApplicationFactory, ApiOAuth2PersonalTokenFactory, AuthUserFactory,
+    BookmarkCollectionFactory, CollectionFactory, MockAddonNodeSettings, NodeFactory,
+    NodeLogFactory, PrivateLinkFactory, ProjectWithAddonFactory, ProjectFactory,
+    RegistrationFactory, UnconfirmedUserFactory, UnregUserFactory, UserFactory, WatchConfigFactory,
+    InstitutionFactory,
+)
+from tests.test_features import requires_search
+from website.util import api_url_for
+
+
+@requires_search
+class TestSearchViews(OsfTestCase):
+
+    def setUp(self):
+        super(TestSearchViews, self).setUp()
+        import website.search.search as search
+        search.delete_all()
+
+        self.project = ProjectFactory(creator=UserFactory(fullname='Robbie Williams'))
+        self.contrib = UserFactory(fullname='Brian May')
+        for i in range(0, 12):
+            UserFactory(fullname='Freddie Mercury{}'.format(i))
+
+    def tearDown(self):
+        super(TestSearchViews, self).tearDown()
+        import website.search.search as search
+        search.delete_all()
+
+    def test_search_contributor(self):
+        url = api_url_for('search_contributor')
+        res = self.app.get(url, {'query': self.contrib.fullname})
+        assert_equal(res.status_code, 200)
+        result = res.json['users']
+        assert_equal(len(result), 1)
+        brian = result[0]
+        assert_equal(brian['fullname'], self.contrib.fullname)
+        assert_in('gravatar_url', brian)
+        assert_equal(brian['registered'], self.contrib.is_registered)
+        assert_equal(brian['active'], self.contrib.is_active)
+
+    def test_search_pagination_default(self):
+        url = api_url_for('search_contributor')
+        res = self.app.get(url, {'query': 'fr'})
+        assert_equal(res.status_code, 200)
+        result = res.json['users']
+        pages = res.json['pages']
+        page = res.json['page']
+        assert_equal(len(result), 5)
+        assert_equal(pages, 3)
+        assert_equal(page, 0)
+
+    def test_search_pagination_default_page_1(self):
+        url = api_url_for('search_contributor')
+        res = self.app.get(url, {'query': 'fr', 'page': 1})
+        assert_equal(res.status_code, 200)
+        result = res.json['users']
+        page = res.json['page']
+        assert_equal(len(result), 5)
+        assert_equal(page, 1)
+
+    def test_search_pagination_default_page_2(self):
+        url = api_url_for('search_contributor')
+        res = self.app.get(url, {'query': 'fr', 'page': 2})
+        assert_equal(res.status_code, 200)
+        result = res.json['users']
+        page = res.json['page']
+        assert_equal(len(result), 2)
+        assert_equal(page, 2)
+
+    def test_search_pagination_smaller_pages(self):
+        url = api_url_for('search_contributor')
+        res = self.app.get(url, {'query': 'fr', 'size': 5})
+        assert_equal(res.status_code, 200)
+        result = res.json['users']
+        pages = res.json['pages']
+        page = res.json['page']
+        assert_equal(len(result), 5)
+        assert_equal(page, 0)
+        assert_equal(pages, 3)
+
+    def test_search_pagination_smaller_pages_page_2(self):
+        url = api_url_for('search_contributor')
+        res = self.app.get(url, {'query': 'fr', 'page': 2, 'size': 5, })
+        assert_equal(res.status_code, 200)
+        result = res.json['users']
+        pages = res.json['pages']
+        page = res.json['page']
+        assert_equal(len(result), 2)
+        assert_equal(page, 2)
+        assert_equal(pages, 3)
+
+    def test_search_projects(self):
+        url = '/search/'
+        res = self.app.get(url, {'q': self.project.title})
+        assert_equal(res.status_code, 200)
+
+
+class TestODMTitleSearch(OsfTestCase):
+    """ Docs from original method:
+    :arg term: The substring of the title.
+    :arg category: Category of the node.
+    :arg isDeleted: yes, no, or either. Either will not add a qualifier for that argument in the search.
+    :arg isFolder: yes, no, or either. Either will not add a qualifier for that argument in the search.
+    :arg isRegistration: yes, no, or either. Either will not add a qualifier for that argument in the search.
+    :arg includePublic: yes or no. Whether the projects listed should include public projects.
+    :arg includeContributed: yes or no. Whether the search should include projects the current user has
+        contributed to.
+    :arg ignoreNode: a list of nodes that should not be included in the search.
+    :return: a list of dictionaries of projects
+    """
+    def setUp(self):
+        super(TestODMTitleSearch, self).setUp()
+
+        self.user = AuthUserFactory()
+        self.user_two = AuthUserFactory()
+        self.project = ProjectFactory(creator=self.user, title="foo")
+        self.project_two = ProjectFactory(creator=self.user_two, title="bar")
+        self.public_project = ProjectFactory(creator=self.user_two, is_public=True, title="baz")
+        self.registration_project = RegistrationFactory(creator=self.user, title="qux")
+        self.folder = CollectionFactory(creator=self.user, title="quux")
+        self.dashboard = BookmarkCollectionFactory(creator=self.user, title="Dashboard")
+        self.url = api_url_for('search_projects_by_title')
+
+    def test_search_projects_by_title(self):
+        res = self.app.get(self.url, {'term': self.project.title}, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_equal(len(res.json), 1)
+        res = self.app.get(self.url,
+                           {
+                               'term': self.public_project.title,
+                               'includePublic': 'yes',
+                               'includeContributed': 'no'
+                           }, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_equal(len(res.json), 1)
+        res = self.app.get(self.url,
+                           {
+                               'term': self.project.title,
+                               'includePublic': 'no',
+                               'includeContributed': 'yes'
+                           }, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_equal(len(res.json), 1)
+        res = self.app.get(self.url,
+                           {
+                               'term': self.project.title,
+                               'includePublic': 'no',
+                               'includeContributed': 'yes',
+                               'isRegistration': 'no'
+                           }, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_equal(len(res.json), 1)
+        res = self.app.get(self.url,
+                           {
+                               'term': self.project.title,
+                               'includePublic': 'yes',
+                               'includeContributed': 'yes',
+                               'isRegistration': 'either'
+                           }, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_equal(len(res.json), 1)
+        res = self.app.get(self.url,
+                           {
+                               'term': self.public_project.title,
+                               'includePublic': 'yes',
+                               'includeContributed': 'yes',
+                               'isRegistration': 'either'
+                           }, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_equal(len(res.json), 1)
+        res = self.app.get(self.url,
+                           {
+                               'term': self.registration_project.title,
+                               'includePublic': 'yes',
+                               'includeContributed': 'yes',
+                               'isRegistration': 'either'
+                           }, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_equal(len(res.json), 2)
+        res = self.app.get(self.url,
+                           {
+                               'term': self.registration_project.title,
+                               'includePublic': 'yes',
+                               'includeContributed': 'yes',
+                               'isRegistration': 'no'
+                           }, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_equal(len(res.json), 1)
+        res = self.app.get(self.url,
+                           {
+                               'term': self.folder.title,
+                               'includePublic': 'yes',
+                               'includeContributed': 'yes',
+                               'isFolder': 'yes'
+                           }, auth=self.user.auth, expect_errors=True)
+        assert_equal(res.status_code, 404)
+        res = self.app.get(self.url,
+                           {
+                               'term': self.folder.title,
+                               'includePublic': 'yes',
+                               'includeContributed': 'yes',
+                               'isFolder': 'no'
+                           }, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_equal(len(res.json), 0)
+        res = self.app.get(self.url,
+                           {
+                               'term': self.dashboard.title,
+                               'includePublic': 'yes',
+                               'includeContributed': 'yes',
+                               'isFolder': 'no'
+                           }, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        assert_equal(len(res.json), 0)
+        res = self.app.get(self.url,
+                           {
+                               'term': self.dashboard.title,
+                               'includePublic': 'yes',
+                               'includeContributed': 'yes',
+                               'isFolder': 'yes'
+                           }, auth=self.user.auth, expect_errors=True)
+        assert_equal(res.status_code, 404)
